entrega 2: comienzo desarrollo algoritmo simplex

+++ src/main/java/domain/DispositivoEstandar.java
@@ -1,11 +1,11 @@
 package domain;
 
 public class DispositivoEstandar extends Dispositivo {
-	public DispositivoEstandar(int dispositivo, String nombre, int consumoXHora) {
+	public DispositivoEstandar(int dispositivo, String nombre, double consumoXHora) {
 		super(dispositivo, nombre, consumoXHora);
 	}
 
-	public int estimarConsumo(int horasUsoXDia) {
+	public double estimarConsumo(int horasUsoXDia) {
 		return this.getConsumoXHora() * horasUsoXDia;
 	}
 }
